\name{HPdrecommend-package}
\alias{HPdrecommend-package}
\alias{HPdrecommend}
\docType{package}
\title{
\packageTitle{HPdrecommend}
}
\description{
\packageDescription{HPdrecommend}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HPdrecommend}
\packageIndices{HPdrecommend}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HPdrecommend}

Maintainer: \packageMaintainer{HPdrecommend}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
